safari-534.53-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.1_Reserved_Words/S7.5.1_A1.3-expected.txt

S7.5.1_A1.3

PASS ReferenceError: Left side of assignment is not a reference.

TEST COMPLETE
